Update routes, add validation and handle controllers

+++ controllers/profileController.js
@@ -3,8 +3,9 @@ const { body, validationResult } = require('express-validator');
 const Profile = require('../models/Profile');
 const Tweet = require('../models/Tweet');
 const User = require('../models/User');
-const startCase = require("lodash");
+const startCase = require('../helpers/startCase');
 const charLengthForProps = require('../helpers/charLengthForProps');
+const validateBirthday = require('../validation/birthday');
 
 exports.getLoggedInUserProfile = async (req, res, next) => {
     try {
@@ -14,7 +15,7 @@ exports.getLoggedInUserProfile = async (req, res, next) => {
         );
 
         if (!profile) {
-            return res.status(404).json({ msg: 'Profile does not exists' });
+            return res.status(404).json({ errors: [{ msg: 'Profile does not exists' }]});
         }
 
         res.json(profile);
@@ -34,7 +35,7 @@ exports.getLoggedInUserProfileWithHomepageTweets = async (req, res, next) => {
             });
 
         if (!profile) {
-            return res.status(404).json({ message: 'Profile Not Found' });
+            return res.status(404).json({ errors: [{ msg: 'Profile Not Found' }]});
         }
         res.json(profile);
     } catch (err) {
@@ -53,7 +54,7 @@ exports.getLoggedInUserProfileWithTweets = async (req, res, next) => {
             });
 
         if (!profile) {
-            return res.status(404).json({ message: 'Profile Not Found' });
+            return res.status(404).json({ errors: [{ msg: 'Profile Not Found' }]});
         }
         res.json(profile);
     } catch (err) {
@@ -86,14 +87,14 @@ exports.getProfileByUserId = async (req, res, next) => {
         ]);
 
         if (!profile) {
-            return res.status(404).json({ msg: 'Profile not found' });
+            return res.status(404).json({ errors: [{ msg: 'Profile not found' }]});
         }
 
         res.json(profile);
     } catch (err) {
         console.error(err.message);
         if (err.kind === 'ObjectId') {
-            return res.status(404).json({ msg: 'Profile not found' });
+            return res.status(404).json({ errors: [{ msg: 'Profile not found' }]});
         }
         next(err);
     }
@@ -107,7 +108,7 @@ exports.getProfileByUsername = async (req, res, next) => {
         if (!user) {
             return res
                 .status(404)
-                .json({ message: 'User with that username does not exists' });
+                .json({ errors: [{ msg: 'User with that username does not exists' }]});
         }
         const profile = await Profile.findOne({ user: user.id }).populate('user', [
             'name',
@@ -117,7 +118,7 @@ exports.getProfileByUsername = async (req, res, next) => {
         if (!profile) {
             return res
                 .status(404)
-                .json({ msg: `Profile for ${username} does not exists` });
+                .json({ errors: [{ msg: `Profile for ${username} does not exists` }]});
         }
 
         res.json(profile);
@@ -135,7 +136,7 @@ exports.getProfileByUsernameWithTweets = async (req, res, next) => {
         if (!user) {
             return res
                 .status(404)
-                .json({ message: 'User with that username does not exists' });
+                .json({ errors: [{ msg: 'User with that username does not exists' }]});
         }
         const profile = await Profile.findOne({ user: user.id })
             .populate('user', ['name', 'username', 'avatar'])
@@ -147,7 +148,7 @@ exports.getProfileByUsernameWithTweets = async (req, res, next) => {
         if (!profile) {
             return res
                 .status(404)
-                .json({ message: 'Profile with that username does not exists' });
+                .json({ errors: [{ msg: 'Profile with that username does not exists' }] });
         }
         res.json(profile);
     } catch (err) {
@@ -158,12 +159,10 @@ exports.getProfileByUsernameWithTweets = async (req, res, next) => {
 
 exports.updateProfile =  async (req, res, next) => {
     const errors = validationResult(req);
-
     if (!errors.isEmpty()) {
-        return res.status(400).json({ errors: errors.array() });
+        return res.status(422).json({ errors: errors.array({ onlyFirstError: true }) });
     }
 
-    // Validation before checks if name exists, if value is empty and if it is alphanumeric, so here I only check if property name in req.body is "something" in order to change it
     try {
         if (req.body['name']) {
             await User.findOneAndUpdate(
@@ -187,9 +186,7 @@ exports.updateProfile =  async (req, res, next) => {
             'location',
             'website',
             'birthday',
-            'avatar',
-            'backgroundPicture',
-            'created'
+            'backgroundPicture'
         ];
 
         standardFields.forEach(field => {
@@ -214,6 +211,11 @@ exports.updateProfile =  async (req, res, next) => {
 
 exports.deleteAccount = async (req, res, next) => {
     try {
+        const user = await User.findById(req.user.id);
+        if (!user) {
+            return res.status(404).json({ errors: [{ msg: 'User Not Found' }]})
+        }
+
         await User.findByIdAndRemove(req.user.id);
         await Profile.findOneAndRemove({ user: req.user.id });
 
@@ -246,39 +248,52 @@ exports.validate = method => {
     switch(method) {
         case 'updateProfile': {
             return [
-                body('bio', `The bio must be between ${charLengthForProps.bio.min} and ${charLengthForProps.bio.max} chars long`)
-                    .optional()
+                body('bio')
+                    .optional({ checkFalsy: true })
                     .trim()
                     .isLength(charLengthForProps.bio)
+                    .withMessage(`The bio must be between ${charLengthForProps.bio.min} and ${charLengthForProps.bio.max} chars long`)
                     .escape(),
-                body('location', `The location must be between ${charLengthForProps.location.min} and ${charLengthForProps.location.max} chars long`)
-                    .optional()
+                body('location')
+                    .optional({ checkFalsy: true })
                     .trim()
-                    .isLength(charLengthForProps.location),
-                body('website', `The website must be between ${charLengthForProps.website.min} and ${charLengthForProps.website.max} chars long`)
-                    .optional()
+                    .isLength(charLengthForProps.location)
+                    .withMessage(`The location must be between ${charLengthForProps.location.min} and ${charLengthForProps.location.max} chars long`),
+                body('website')
+                    .optional({ checkFalsy: true })
+                    .trim()
                     .isLength(charLengthForProps.website)
+                    .withMessage(`The website must be between ${charLengthForProps.website.min} and ${charLengthForProps.website.max} chars long`)
                     .isURL()
                     .withMessage('Website must be a valid URL'),
                 body('birthday', 'Birthday must be a valid date')
-                    .optional()
-                    .toDate(),
+                    .optional({ checkFalsy: true })
+                    .custom(validateBirthday),
                 body('avatar', 'Avatar must be a URL')
-                    .optional()
+                    .optional({ checkFalsy: true })
                     .isURL(),
                 body('backgroundPicture', 'Background picture must be a URL')
-                    .optional()
+                    .optional({ checkFalsy: true })
                     .isURL(),
-                body('name', `The name must be between ${charLengthForProps.name.min} and ${charLengthForProps.name.max} chars long`)
+                body('name', 'The name must be a valid name')
                     .optional()
                     .isLength(charLengthForProps.name)
-                    .isAlpha()
-                    .withMessage('The name must be a valid name'),
-                body('username', `The username must be between ${charLengthForProps.username.min} and ${charLengthForProps.username.max} chars long`)
+                    .withMessage(`The name must be between ${charLengthForProps.name.min} and ${charLengthForProps.name.max} chars long`),
+                body('username', 'The username must be a valid username')
                     .optional()
                     .isLength(charLengthForProps.username)
-                    .isAlpha()
-                    .withMessage('The username must be a valid username')
+                    .withMessage(`The username must be between ${charLengthForProps.username.min} and ${charLengthForProps.username.max} chars long`)
+                    .custom(value => {
+                        const username = value.toString().split(' ').join('');
+                        return User.findOne({ username }).then(user => {
+                            if (user) {
+                                return Promise.reject('username already in use')
+                            }
+                        });
+                    })
+                    .customSanitizer(value => {
+                        return value.split(' ').join('')
+                    })
             ]
         }
     }
